Move software settings handling code to a module
