+++ rust/agama-lib/src/proxies.rs
@@ -140,93 +140,6 @@ trait Questions1 {
 }
 
 #[dbus_proxy(
-    interface = "org.opensuse.Agama.Software1",
-    default_service = "org.opensuse.Agama.Software1",
-    default_path = "/org/opensuse/Agama/Software1"
-)]
-trait Software1 {
-    /// Finish method
-    fn finish(&self) -> zbus::Result<()>;
-
-    /// Install method
-    fn install(&self) -> zbus::Result<()>;
-
-    /// IsPackageInstalled method
-    fn is_package_installed(&self, name: &str) -> zbus::Result<bool>;
-
-    /// Probe method
-    fn probe(&self) -> zbus::Result<()>;
-
-    /// Propose method
-    fn propose(&self) -> zbus::Result<()>;
-
-    /// ProvisionSelected method
-    fn provision_selected(&self, provision: &str) -> zbus::Result<bool>;
-
-    /// ProvisionsSelected method
-    fn provisions_selected(&self, provisions: &[&str]) -> zbus::Result<Vec<bool>>;
-
-    /// SelectProduct method
-    fn select_product(&self, product_id: &str) -> zbus::Result<()>;
-
-    /// UsedDiskSpace method
-    fn used_disk_space(&self) -> zbus::Result<String>;
-
-    /// AvailableBaseProducts property
-    #[dbus_proxy(property)]
-    fn available_base_products(
-        &self,
-    ) -> zbus::Result<
-        Vec<(
-            String,
-            String,
-            std::collections::HashMap<String, zbus::zvariant::OwnedValue>,
-        )>,
-    >;
-
-    /// SelectedBaseProduct property
-    #[dbus_proxy(property)]
-    fn selected_base_product(&self) -> zbus::Result<String>;
-}
-
-#[dbus_proxy(
-    interface = "org.opensuse.Agama.Software1.Proposal",
-    default_service = "org.opensuse.Agama.Software1",
-    default_path = "/org/opensuse/Agama/Software1/Proposal"
-)]
-trait SoftwareProposal {
-    /// AddResolvables method
-    fn add_resolvables(
-        &self,
-        id: &str,
-        r#type: u8,
-        resolvables: &[&str],
-        optional: bool,
-    ) -> zbus::Result<()>;
-
-    /// GetResolvables method
-    fn get_resolvables(&self, id: &str, r#type: u8, optional: bool) -> zbus::Result<Vec<String>>;
-
-    /// RemoveResolvables method
-    fn remove_resolvables(
-        &self,
-        id: &str,
-        r#type: u8,
-        resolvables: &[&str],
-        optional: bool,
-    ) -> zbus::Result<()>;
-
-    /// SetResolvables method
-    fn set_resolvables(
-        &self,
-        id: &str,
-        r#type: u8,
-        resolvables: &[&str],
-        optional: bool,
-    ) -> zbus::Result<()>;
-}
-
-#[dbus_proxy(
     interface = "org.opensuse.Agama.Users1",
     default_service = "org.opensuse.Agama.Users1",
     default_path = "/org/opensuse/Agama/Users1"
